feat(spacy): reformat retrain spacy code
